refactor: 입출력 콘솔 클래스 구현에 따라 앱 컨트롤러 코드 및 패키지 변경
